Changed tabs to spaces in split_op, implemented Grid variable map
We need to think about how we are going to store grid-specific variables and
the necessary functions for the Wave class.

Also think about ways to overload the return so the val() function will return
stuff regardless of whether we are storing ints or doubles.

Also think about how we want to retrieve the data. Write a macro.